moved functions and interfaces to separate files for easier readability
